requirements.txt
pycryptocome
dbus-python
gatt
matplotlib
numpy